+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For a full diff see [`1.3.1...master`][1.3.1...master]
 
+## Changed
+
+* Started using `ergebnis/composer-json-normalizer` instead of `localheinz/composer-json-normalizer`, `ergebnis/json-normalizer` instead of `localheinz/json-normalizer`, and `ergebnis/json-printer` instead of `localheinz/json-printer` ([#261]), by [@localheinz]
+
 ### Fixed
 
 * Dropped support for PHP 7.1 ([#235]), by [@localheinz]
@@ -247,6 +251,7 @@ For a full diff see [`81bc3a8...0.1.0`][81bc3a8...0.1.0].
 [#190]: https://github.com/localheinz/composer-normalize/pull/190
 [#207]: https://github.com/localheinz/composer-normalize/pull/207
 [#235]: https://github.com/localheinz/composer-normalize/pull/235
+[#261]: https://github.com/localheinz/composer-normalize/pull/261
 
 [@localheinz]: https://github.com/localheinz
 [@svenluijten]: https://github.com/svenluijten